+++ npm/ng-packs/packages/setting-management/config/src/services/index.ts
@@ -1,1 +1,1 @@
-export * from './settings-tab.service';
+export * from './settings-tabs.service';
