+++ npm/ng-packs/packages/setting-management/config/src/services/settings-tabs.service.ts
